% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcausation.R
\name{pcausation}
\alias{pcausation}
\title{Semiparametric estimation of the probability of causation}
\usage{
pcausation(y, a, x, xtest, nsplits = 2, start.list = c(rep(0,
  ncol(x))), tracetf = FALSE, printres = TRUE, showprogress = TRUE,
  sl.lib = c("SL.earth", "SL.gam", "SL.glm", "SL.glm.interaction",
  "SL.mean", "SL.ranger", "SL.rpart"))
}
\arguments{
\item{y}{A binary vector, of 0s and 1s, representing the outcome (e.g. of an intervention).}

\item{a}{A binary vector, of 0s and 1s, representing the exposure (e.g. to a treatment).}

\item{x}{A data frame of covariates x1...xn.}

\item{xtest}{A data frame of covariates x1...xn at which to evaluate the probability of causation. Could be same as x.}

\item{nsplits}{How many splits to use for cross-validation. If none are desired, use 1.}

\item{start.list}{Which values to use in starting optimization. Default is vector of 0s.}

\item{tracetf}{Do you want to show the trace, i.e., progress, of the optimization?}

\item{printres}{Do you want to print the table of results?}

\item{showprogress}{Do you want to see a progress bar while the estimation is performed?}

\item{sl.lib}{Library of algorithms to be given to superlearner. Default is a mixture.}
}
\value{
A table of estimates of the probability of causation with confidence intervals, and other parameters.
}
\description{
Takes in values of binary outcome y, binary exposure a, and covariates x1...xn, and outputs prob of causation and other parameters.
}
